Handle None TypeError on utc_offset
If you are running on a system (like Linux) using UTC as the local timezone then ipapi may return None as the utc_offset.  This results in a TypeError when calculating the utc_offset value for later use.  This fixes it by checking to see if it is None in advance of doing the math and setting to 0 if it is.