Telegram Bot
I added the possibility to be up to date from a telegram bot that send messagge when start, every time it ends an account with the daily earned points and total points in the account, when the script ends with points sum of all account.

If the settings are not configured the script can run anyway, texting that the file telegram.json were create like the accounts.json file.

If you like the idea i can write the instruction in the README.md.